Implement closing logic according to spec
The current version of the code didn't actually transition from Established to
Closed properly on the peer that initiated the close.  Only receiving
CONNECTION_CLOSE caused the connection to be closed.  Rather than dance around
the edges of this, I decided to implement the spec.

Part of that includes implementing RTO.