Stop treating RST_STREAM as EOF
In hq, it's important to know when the stream is reset as opposed to ending naturally.  You need to know this so that you can acknowledge any potentially unread header blocks.

This changes the code so that resets cause subsequent reads to immediately fail with a recognizable error code.  Receiving RST_STREAM stops further reading.  It already caused existing data to be discarded.  Finally, when a read fails on a reset stream, the state moves to the previously unused ResetRead state.